Feature/rutas

+++ src/main/java/org/example/Repository/RoleR.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RoleR extends JpaRepository<Role, Long> {
+public interface RoleR extends JpaRepository<Role, Integer> {
     Optional<Role> findByRol(String rol);
 }
